Fixed warnings when size_t != unsigned long long int

+++ src/serverconfig.cxx
@@ -60,7 +60,7 @@ void ServerConfig::readIcon()
   if (fs::exists(iconpath))
   {
     ByteBuffer buffer;
-    buffer.resize(fs::file_size(iconpath));
+    buffer.resize(static_cast<std::size_t>(fs::file_size(iconpath)));
 
     {
       std::ifstream file(iconpath);
